+++ tests/wires/test_contour2d.py
@@ -1,7 +1,7 @@
 import unittest
 
 import volmdlr
-from volmdlr.models.contours import Contour2d_1
+from volmdlr.models.contours import contour2d_1, contour2d_2
 from volmdlr import wires, edges
 
 
@@ -13,9 +13,14 @@ class TestContour2D(unittest.TestCase):
         point2 = volmdlr.Point2D(0.02, 0.02)
         self.assertTrue(self.contour1.point_belongs(point1))
         self.assertFalse(self.contour1.point_belongs(point2))
+
         point3 = volmdlr.Point2D(0, 0.013)
-        self.assertTrue(Contour2d_1.point_belongs(point3))
-        self.assertFalse(Contour2d_1.point_belongs(point1))
+        self.assertTrue(contour2d_1.point_belongs(point3))
+        self.assertFalse(contour2d_1.point_belongs(point1))
+
+        point4 = volmdlr.Point2D(0.745, 0.0685)
+        self.assertTrue(contour2d_2.point_belongs(point4))
+
 
     def test_cut_by_wire(self):
         pass